requirements.txt
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.2
Markdown==3.7
sqlparse==0.5.3
validate-docbr==1.10.0
